use TameAutoClosable and move DataCollectors to smith
