+++ plugins/org.eclipse.n4js.transpiler.es/META-INF/MANIFEST.MF
@@ -7,7 +7,6 @@ Bundle-Vendor: %providerName
 Bundle-RequiredExecutionEnvironment: JavaSE-11
 Require-Bundle: org.eclipse.n4js,
  org.eclipse.n4js.transpiler,
- org.eclipse.n4js.utils,
  org.apache.log4j,
  org.eclipse.n4js.ts,
  org.eclipse.n4js.ts.model,
