+++ plugins/org.eclipse.n4js.transpiler.es/src/org/eclipse/n4js/transpiler/es/EcmaScriptSubGenerator.java
@@ -27,9 +27,9 @@ import org.eclipse.n4js.n4JS.Script;
 import org.eclipse.n4js.projectModel.IN4JSProject;
 import org.eclipse.n4js.resource.N4JSResource;
 import org.eclipse.n4js.smith.Measurement;
+import org.eclipse.n4js.smith.N4JSDataCollectors;
 import org.eclipse.n4js.transpiler.AbstractTranspiler;
 import org.eclipse.n4js.transpiler.AbstractTranspiler.SourceMapInfo;
-import org.eclipse.n4js.utils.N4JSDataCollectors;
 import org.eclipse.xtext.generator.IFileSystemAccess;
 import org.eclipse.xtext.generator.OutputConfiguration;
 import org.eclipse.xtext.util.CancelIndicator;
