+++ plugins/org.eclipse.n4js.transpiler/src/org/eclipse/n4js/transpiler/AbstractTranspiler.java
@@ -19,10 +19,10 @@ import org.eclipse.n4js.resource.N4JSResource;
 import org.eclipse.n4js.smith.DataCollector;
 import org.eclipse.n4js.smith.DataCollectors;
 import org.eclipse.n4js.smith.Measurement;
+import org.eclipse.n4js.smith.N4JSDataCollectors;
 import org.eclipse.n4js.transpiler.print.PrettyPrinter;
 import org.eclipse.n4js.transpiler.utils.TranspilerDebugUtils;
 import org.eclipse.n4js.ts.types.TModule;
-import org.eclipse.n4js.utils.N4JSDataCollectors;
 import org.eclipse.n4js.utils.ResourceNameComputer;
 import org.eclipse.n4js.utils.di.scopes.ScopeManager;
 import org.eclipse.n4js.utils.di.scopes.TransformationScoped;
