+++ plugins/org.eclipse.n4js.utils/META-INF/MANIFEST.MF
@@ -20,8 +20,7 @@ Require-Bundle: org.eclipse.n4js.releng.utils;resolution:=optional,
  com.jcraft.jsch;bundle-version="0.1.53",
  org.slf4j.api,
  org.apache.commons.logging;resolution:=optional,
- org.eclipse.equinox.common;resolution:=optional,
- org.eclipse.n4js.smith
+ org.eclipse.equinox.common;resolution:=optional
 Export-Package: org.eclipse.n4js.fileextensions,
  org.eclipse.n4js.utils,
  org.eclipse.n4js.utils.beans,
