+++ plugins/org.eclipse.n4js.utils/src/org/eclipse/n4js/utils/TameAutoClosable.java
@@ -1,0 +1,21 @@
+/**
+ * Copyright (c) 2019 NumberFour AG.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *   NumberFour AG - Initial API and implementation
+ */
+package org.eclipse.n4js.utils;
+
+/**
+ * Same as {@link AutoCloseable} but does not throw {@link Exception}s.
+ */
+public interface TameAutoClosable extends AutoCloseable {
+
+	@Override
+	void close();
+
+}
